- add service layer template - add default exception - add image storage service template

+++ user-service/src/main/java/ru/skillbox/userservice/model/dto/AccountSearchDto.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 @Data
 public class AccountSearchDto {
-    List<Integer> ids;
+    List<Long> ids;
     String author;
     String firstName;
     String lastName;
